Moved job names from static strings to string references stored in constants.ts
