+++ src/database/constants.ts
@@ -1,0 +1,2 @@
+export const GET_ATTACK_OBJECTS_JOB_TOKEN = "GET_ATTACK_OBJECTS";
+export const GET_STIX_COLLECTIONS_JOB_TOKEN = "GET_STIX_COLLECTIONS";
